terminal-client/pyproject.toml
[project]
name = "narrator-client"
version = "0.1.0"
description = "Narrator MCP Client - Bridge for running commands with narration"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pyaudio>=0.2.13",
    "pydub>=0.25.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

